+++ cmake/QuillUtils.cmake
@@ -34,7 +34,7 @@ function(set_common_compile_options target_name)
 
     target_compile_options(${target_name} ${COMPILE_OPTIONS_VISIBILITY}
             $<$<OR:$<CXX_COMPILER_ID:Clang>,$<CXX_COMPILER_ID:AppleClang>,$<CXX_COMPILER_ID:GNU>>:
-            -Wall -Wextra -pedantic -Werror>
+            -Wall -Wextra -pedantic -Werror -Wredundant-decls>
             $<$<CXX_COMPILER_ID:MSVC>:/bigobj /WX /W4 /wd4324 /wd4996>)
 
     if (QUILL_NO_EXCEPTIONS)