+++ CHANGELOG.md
@@ -80,6 +80,10 @@
 - [v1.1.0](#v110)
 - [v1.0.0](#v100)
 
+## TBD
+
+- Suppress `-Wredundant-decls` warning in GCC builds.
+
 ## v7.5.0
 
 - In previous versions, logging on Windows automatically included `windows.h` in all components. The frontend will no